requirements.txt
amqp==2.1.4
asn1crypto==0.22.0
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==1.9
Flask==0.12.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.0.2
MarkupSafe==1.0
pycparser==2.17
pyOpenSSL==17.1.0
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
vine==1.1.3
Werkzeug==0.12.2
